Create the log file only when it does not exist
The main log file is overwritten every time any appjail's rc script is
called, to avoid this, the log file is created only when it does not
exist.

+++ etc/rc.d/appjail-natnet.sh
@@ -30,7 +30,9 @@ status_cmd="${name}_status"
 
 precmd()
 {
-	install -m 640 /dev/null "${appjail_natnet_logfile}"
+	if [ ! -f "${appjail_natnet_logfile}" ]; then
+		install -m 640 /dev/null "${appjail_natnet_logfile}"
+	fi
 }
 
 _show_logfile()
